refactor(user): user.view.ts를 역할별 3개 파일로 분리
- src/user/view/ 서브디렉토리 신설
- user.view.ts — 공통 타입/상수 + 회원가입 모달 + 내 정보 모달
- user-admin.view.ts (UserAdminView) — 승인 대기 · 유저 목록 · 편집 모달
- user-class-rep.view.ts (UserClassRepView) — 반 승인 대기 · 반원 목록 모달
- UserController import 경로 및 호출 클래스 업데이트
